+++ interface/mmal/vc/mmal_vc_diag.c
@@ -379,14 +379,14 @@ static int do_eventlog(int argc, const char **argv)
    }
    if (log.size != sizeof(log))
    {
-      fprintf(stderr,"MMAL Log size mismatch (got %d, expected %d)\n",
-              log.size, (int)sizeof(log));
+      fprintf(stderr,"MMAL Log size mismatch (got %d, expected %ld)\n",
+              log.size, (long)sizeof(log));
       goto fail;
    }
    if (log.elemsize != sizeof(MMAL_DBG_ENTRY_T))
    {
-      fprintf(stderr,"MMAL log element size mismatch (got %d, expected %d)\n",
-              log.elemsize, (int)sizeof(MMAL_DBG_ENTRY_T));
+      fprintf(stderr,"MMAL log element size mismatch (got %d, expected %ld)\n",
+              log.elemsize, (long)sizeof(MMAL_DBG_ENTRY_T));
       goto fail;
    }
